Move wf chain tests to their own module.
